chore(build): replace binary favicon with SVG
